server/requirements.txt
Flask
Flask-Bcrypt
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
python-dotenv
gunicorn
psycopg2-binary